--- src/dataCompose/index.js
@@ -1,1 +1,0 @@
-export * from './dataCompose';
